% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Test_and_Extras.R
\name{diego_pv}
\alias{diego_pv}
\title{Estas eran mas que todo para el Poster, para poder ensenar los graficos.
Enfocarse mas en las weight_functions, mira el papel, dejar la value function
como esta y documentarla, esto solo como extra, ignore o play.}
\usage{
diego_pv(dataset, userid = NULL, attr = NULL, rounds = NULL,
  refps = NULL, cost_ids = NULL, weight = NULL, gainm = TRUE)
}
\description{
as dos funciones solo son para poder sacar la value function mas simple,
simple weighted additive function, gain_matrix sin normalizar y multiplicarlo
por los pesos, no es igual que el paper pero podria ser util, decide quickly.
}

