% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeightFunctions.R
\name{differenceToIdeal}
\alias{differenceToIdeal}
\title{Calculates decision weights using the 'objective approach'}
\usage{
differenceToIdeal(normalizedMatrix, attr)
}
\arguments{
\item{normalizedMatrix}{a numeric matrix. If indeed normalized it should only
contain values between \code{0} and \code{1}.}

\item{attr}{attributes IDs, vector of integer numbers corresponding to the
attributes you desire to use; attr are assumed to be 1-indexed. Here it
represents the number of columns of the input matrix.}
}
\value{
a decision weight (numeric vector with a sum of 1)
}
\description{
This function is used as the second step in
\code{\link{weight.differenceToIdeal}} for calculating a decision weight for
each \code{\link{attr}} in the decision matrix. The methodology of the
'objective approach' for determining the weights is given by references [1]
and [2]. See Details.
}
\details{
The sum of the output of this functions should always equal 1.

It measures the distance of each column value against the best value
  for a given attribute. A smaller difference should mean that for that attribute a
  high value was \emph{consistently} taking in consideration, thus resulting
  in a higher weight. Two unintended consequences are: 1. matrices with one
  row will result in the same weight for all columns and 2. for an attribute
  where the value does not change at all (even if it's a low value) the
  function will reward it somewhat disproportionately.
}
\examples{
#Runnable
differenceToIdeal(matrix(c(1.0, 0.85, 0.42, 0, 0.5, 0, 1, 0.7), 4, 2), attr=1:4)
weights <- differenceToIdeal(matrix(c(1.0, 0.85, 0.42, 0, 0.5, 0, 1, 0.7), 2, 4), attr=1:4)
sum(weights) should return 1

}
\references{
[1]Ma, J., Fan, Z. P., & Huang, L. H. (1999). A subjective and
  objective integrated approach to determine attribute weights. European
  journal of operational research, 112(2), 397-404.

  [2] Fan, Z. P. (1996). Complicated multiple attribute decision making:
  theory and applications (Doctoral dissertation, Ph. D. Dissertation,
  North-eastern University, Shenyang, PRC).
}

