% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRPFunctions.R
\name{dualLossAversionFun}
\alias{dualLossAversionFun}
\title{Takes a gain-loss matrix and outputs a value matrix}
\usage{
dualLossAversionFun(gainLossMatrix, decMatrix, lambda = 2.25, delta = 0.8)
}
\arguments{
\item{gainLossMatrix}{a gain-loss matrix, such as the output from
\code{\link{dualGainLossFunction}}. Format: \code{attr} are the columns and
\code{rounds} as rows. (Numerical)}

\item{decMatrix}{the decision matrix.}

\item{lambda}{numeric - parameter of loss aversion for the value function as
given by reference[1]. Default value is 2.25 as given by [2].}

\item{delta}{numeric - expresses the relative importance of the aspiration
level to other factors. Default is 0.8 and it should satisfy \code{0 <
delta <1}.}
}
\value{
a value matrix with equal dimensions as \code{gainLossMatrix}
}
\description{
This function is called by \code{\link{smallerThanZero}} for its third step.
It runs the 'loss-aversion' function over the values of the gain-loss matrix,
and a consumption utility function over the values of the second parameter,
as shown in [1]. Both input matrices must have the same dimensions.
Altogether it returns a value matrix.
}
\examples{
#Runnable
dualLossAversionFun(matrix(16:31, 4, 4), matrix(51:36, 4, 4))
dualLossAversionFun(matrix(16:31, 4, 4), matrix(1:16, 4, 4), lambda = 5, delta = 0.7)
dualLossAversionFun(matrix(1:100, 5, 20, byrow= T), matrix(150:51, 5, 20),delta = 0.3)

#Not runnable yet
dualGainLossFunction(aGainLossMatrix, itsDecisionMatrix)

}
\references{
[1] Golman, R., & Loewenstein, G. (2011). Explaining Nonconvex
  Preferences with Aspirational and Status Quo Reference Dependence. Mimeo,
  Carnegie Mellon University.

  [2] Tversky, A., & Kahneman, D. (1992). Advances in prospect theory:
  Cumulative representation of uncertainty. Journal of Risk and uncertainty,
  5(4), 297-323.
}

