% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRPFunctions.R
\name{dualValueMatrix.oneAttr}
\alias{dualValueMatrix.oneAttr}
\title{Returns a Value Matrix using two reference points (one attribute only)}
\usage{
dualValueMatrix.oneAttr(dataset, userid = NULL, attr = NULL,
  rounds = NULL, cost_ids = NULL, dual.refps = c(sq = NA, g = NA),
  lambda = 2.25, delta = 0.8, consumption_fun = NULL)
}
\arguments{
\item{dataset}{data.frame with the user generated data from a product
configurator. See \code{decisionMatrix} for specifications of the dataset.}

\item{userid}{a vector of integers that gives the information of which users
the matrix should be calculated. Vectorised.}

\item{attr}{attributes ID, \emph{one integer} corresponding to the attribute
you desire to use; \code{attr} are assumed to be 1-indexed.}

\item{rounds}{integer vector or text option. Which steps of the configuration
process should be shown? Defaults are first and last step. Text options are
\code{all, first, last}.}

\item{cost_ids}{argument used to convert selected cost attributes into benefit
attributes. Cost attribute means that weith a lower value, the user is
better off than with a higher value (e.g. price). Default assumes all
attributes are of benefit type (higher amount is better).}

\item{dual.refps}{numeric vector - two numbers indicating the status-quo and the
aspiration level(goal) for the given attributes. Status-quo should always be
the first input. Contrary to \code{\link{dualValueMatrix}}, this function
also allows for aspiration levels to be smaller than the status-quo (g < sq)
[1].}

\item{lambda}{numeric - parameter of loss aversion for the value function as
given by [1]. Default value is 2.25 as in [2] and should be \code{lambda >
1}.}

\item{delta}{numeric - expresses the relative importance of the aspiration
level to other factors. Default is 0.8 and it should satisfy \code{0 < delta
<1}.}
}
\value{
a list of value matrices with one attribute for each user.
}
\description{
This function is a more basic function than \code{\link{dualValueMatrix}}. This
function is based on the value function of the dual-reference point model
(dual-rp) [1]. It first builds a desicion matrix for each user and then applys
the drp-utility function over each value using \code{\link{smallerThanZero}}.
The dual-rp value function first creates a gain-loss matrix based on the two
reference points. It then outputs the value of each gain/loss based on the
loss aversion (lambda) and the relative importance of the goal (delta).
}
\details{
This function does the same as \code{\link{dualValueMatrix}} but only
 for one attribute, for more details please see the mentioned function.

 Note: When converting a cost attribute to a benefit attribute its two
 reference points change as well, enter the unconverted refps, the function
 transforms them automatically when it detects a \code{cost_ids  != NULL}
}
\examples{
dualValueMatrix.oneAttr(myData, 10, attr = 3, dual.refps = c(1, 3.5))
dualValueMatrix.oneAttr(myData, userid= 60, rounds= "all", attr = 1, dual.refps = c(1.5, 2.5)
dualValueMatrix.oneAttr(myData, 10, attr=4, dual.refps = c(0.17, -0.10), cost_ids = 4) # Note for cost_ids SQ > G

# Return an error, 1.Too many attributes or 2. none entered
dualValueMatrix.oneAttr(keyboard_data, 8:9 , attr = c(10,12,14,16), dual.refps = c(100, 150))
dualValueMatrix.oneAttr(data1, 2) # 2. No attribute entered

}
\references{
[1] Golman, R., & Loewenstein, G. (2011). Explaining Nonconvex
  Preferences with Aspirational and Status Quo Reference Dependence. Mimeo,
  Carnegie Mellon University.

  [2] Tversky, A., & Kahneman, D. (1992). Advances in prospect theory:
  Cumulative representation of uncertainty. Journal of Risk and uncertainty,
  5(4), 297-323.
}

