% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PT-Functions.R
\name{gainLoss}
\alias{gainLoss}
\title{Merges gain and loss matrices}
\usage{
gainLoss(dataset, userid = NULL, attr = NULL, rounds = NULL,
  refps = NULL, cost_ids = NULL)
}
\arguments{
\item{refps}{a list of numeric vectors, one for each user. Reference Points:
each point corresponds to one attribute, therefore the amount of attributes
and of refps entered, should be equal. Default assumes the refps as the
default values of the initial product configuration for each user. You may
fully or partially enter your own reference points, check below for more
info.}
}
\value{
a list of combined matrices.
}
\description{
The matrices are composed by row-binded gain matrices (on top) and loss
matrices (below). This is more of an auxiliary functions, since it is easiert
to work with united matrices. Rows and columns are unnamed. The output of this
function is further used by \code{\link{norm.gainLoss}}.
}
\details{
\code{dataset} We assume the input data.frame has following columns
  usid = User IDs, round = integers indicating which round the user is in
  (0-index works best for 'round'), atid = integer column for referring the
  attribute ID (1 indexed), selected = numeric value of the attribute for a
  specific, given round, selectable = amount of options the user can chose at
  a given round, with the current configuration. This is a necessary
  parameter.

  \code{dataset} We assume the input data.frame has following columns usid =
  User IDs, round = integers indicating which round the user is in (0-index
  works best for 'round'), atid = integer column for referring the attribute
  ID (1 indexed), selected = numeric value of the attribute for a specific,
  given round, selectable = amount of options the user can chose at a given
  round, with the current configuration. This is a necessary parameter.

  \code{userid} is a necessary parameter, without it you'll get a warning.
  Default is NULL.

  \code{attr} Default calculates with all attributes. Attributes are
  automatically read from provided table, it is important you always provide
  the complete dataset so that the package functions properly. Moreover the
  attributes will not be sorted. Output columns are returned in the ordered
  they were inputed.

  \code{rounds} If you need to compute different rounds for each user you
  enter, this argument accepts a list of integer vectors indicating which
  rounds should be used for each user. The function does not read names, it
  works in the order the list was given.

  \code{refps} If you only want to see the results for one attribute you may
  enter only a couple of reference points but you have to tell the function
  which attributes you want to use those referene points for. So the amount
  of attr and of refps should be the same. Moreover the functions always
  orders de attr, so be sure to input the reference point also in an
  ascending order corresponding to their attributes.
}
\examples{
#Not runnable yet
gainLoss(pc_config_data, 11:12)
gainLoss(my_data, userid = 10:100)
gainLoss(monitor_data, c(50,51), rounds = "last", refps = c(0.1,0.3,0.4,0.5), cost_ids = 3)
gainLoss(data1, 40, attr = 1)
gainLoss(data, 3, attr = c(1,2,3,4))
}

