% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetFunctions.R
\name{get_attr_values}
\alias{get_attr_values}
\title{All unique values for a given attribute}
\usage{
get_attr_values(dataset, attrid = NULL)
}
\arguments{
\item{dataset}{complete data.frame with a \code{atid} column referring to the
attribute ID.}

\item{attrid}{the ID of the attribute you want to know all unique values of.}
}
\value{
Unique values for the given attribute, applies for all users.
}
\description{
Returns all unique values that can be found in the given attribute, i.e. all
values that exist for \code{attrid} for all users.
}
\examples{
get_attr_values(example_data, attrid=1) # Example return: [1] 0 3 2 1 meaning that category only has 4 possible values in the product configutrator, or at least those were the ones the users clicked on
get_attr_values(other_data, 100) #Possible catched error: The attrID you specified is not contain in your data.

}
\seealso{
Other get functions: \code{\link{benefitToCostAttr}},
  \code{\link{get_all_default_rps}},
  \code{\link{get_all_userids}},
  \code{\link{get_attrs_ID}},
  \code{\link{get_normalized_vec}},
  \code{\link{get_rounds_by_ID}},
  \code{\link{get_table_by_ID}}
}

