% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/weight_functions.R
\name{get_attr_weight}
\alias{get_attr_weight}
\title{}
\usage{
get_attr_weight(dataset)
}
\arguments{
\item{rel_frequeny}{logical. Relative frequency, if \code{rel_frequency = TRUE}, the function ignores the refps argument
       and calculates the aspiration level for each attribute as the relative frequency the user made for that attribute.

###to consider, defaults considers all your attributes in your table and calculates with relative frequency of the attributes
if no other weights are given. ##no userid because weights independent from user}
}
\description{

}

