% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/decision_matrix.R, R/get_functions.R
\name{get_attrs_ID}
\alias{get_attrs_ID}
\title{Create the decision matrix (needed Qj and Pi)}
\usage{
get_attrs_ID(x)

get_attrs_ID(x)
}
\arguments{
\item{x}{the data set (table) you want to analize and separate in smaller data sets.}

\item{user_id}{an integer, that identifies which rows of the whole given dataset (x) should be extracted.}

\item{x}{the data set (table) you want to analize and separate in smaller data sets.}

\item{user_id}{an integer, that identifies which rows of the whole given dataset (x) should be extracted.}
}
\value{
A data frame with the same columns as the original, but only with rows about the provided \code{user_id}.

A data frame with the same columns as the original, but only with rows about the provided \code{user_id}.
}
\description{
Description

Description
}
\examples{
get_table_by_ID(camera2_config_data, 9)
get_table_by_ID(camera2_config_data, 9)
}

