% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get_rounds_by_ID}
\alias{get_rounds_by_ID}
\title{Get the amount of rounds (clicks) for one user}
\usage{
get_rounds_by_ID(x, userid = NULL)
}
\arguments{
\item{x}{the complete data.frame. You can also input the data.frame specific for one user, such as the one generated with \code{\link{get_table_by_ID}}. For more users at once see \code{\link{powerful_function}} or manually.#'}

\item{userid}{an integer, that identifies which rows (from which user) of the whole given dataset should be extracted.}
}
\value{
An integer vector representing the index of each round/click and therefore the total amount of rounds for the selected user.
}
\description{
Allows to calculate the amount of times a user interacted with the product configurator, i.e. amount of clicks for the given user ID.
The given data.frame should signalize with a column named 'round', which should only contain numbers.
}
\examples{
get_rounds_by_ID(example_data, userid = 6) # Example return: [1] 0 1 2 3 4 5 6 7 8
get_rounds_by_ID(as.data.frame(matrix_full), 1200) # If userid given not found in the given data(x) function returns: integer(0).
}
\seealso{
\code{\link{powerful_function}} allows you to run this and any other function in this package for more than one user.
}

