% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_functions.R, R/get_table_by_ID.R
\name{get_table_by_ID}
\alias{get_table_by_ID}
\title{Extract the part of the provided dataset that corresponds to a user id.}
\usage{
get_table_by_ID(x, user_id)

get_table_by_ID(x, user_id)
}
\arguments{
\item{x}{the data set (table) you want to analize and separate in smaller data sets.}

\item{user_id}{an integer, that identifies which rows of the whole given dataset (x) should be extracted.}

\item{x}{the data set (table) you want to analize and separate in smaller data sets.}

\item{user_id}{an integer, that identifies which rows of the whole given dataset (x) should be extracted.}
}
\value{
A data frame with the same columns as the original, but only with rows about the provided \code{user_id}.

A data frame with the same columns as the original, but only with rows about the provided \code{user_id}.
}
\description{
We assume the given dataset was generated by data from a product configuration,
and therefore has as one column the user id. Other columns may include, but not limited to, attribute IDs, round,
selected attribute and selectable attribute. With user_id you select the rows, all given columns of x remain intact.

We assume the given dataset was generated by data from a product configuration,
and therefore has as one column the user id. Other columns may include, but not limited to, attribute IDs, round,
selected attribute and selectable attribute. With user_id you select the rows, all given columns of x remain intact.
}
\examples{
get_table_by_ID(camera2_config_data, 9)
get_table_by_ID(camera2_config_data, 9)
}

