% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetFunctions.R
\name{getAllUserIds}
\alias{getAllUserIds}
\title{Get all user IDs}
\usage{
getAllUserIds(dataset)
}
\arguments{
\item{dataset}{the data.frame you want to input.}
}
\value{
a vector of integers with all user IDs. IDs are returned in the order
  they appear on the table.
}
\description{
Allows you to get all the user IDs in your data. It assumes you have a column
\code{usid} that identifies each user's information for each round with its
corresponding user ID. Returns IDS unordered, i.e. in the order they appear
on the table.
}
\examples{
getAllUserIds(example_data) #Example return: [1] 1 2 3 4 5
getAllUserIds(data) #Example return: [1] 6 9 10 8 6 2 4

}
\references{
ProductConfig Github page:
  https://github.com/avilesd/productConfig
}
\seealso{
Other GetFunctions: \code{\link{benefitToCostAttr}},
  \code{\link{getAttrValues}},
  \code{\link{getDefaultRefps}},
  \code{\link{getRoundsById}}, \code{\link{getTableById}},
  \code{\link{get_attrs_ID}}, \code{\link{normalize}}
}

