% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetFunctions.R
\name{getDefaultRefps}
\alias{getDefaultRefps}
\title{Get Default Reference Points for selected user}
\usage{
getDefaultRefps(dataset, userid = NULL)
}
\arguments{
\item{dataset}{the data.frame you want to input. For more details about which information
the dataset should have, refer to \code{\link{decisionMatrix}}.}

\item{userid}{an integer vector, that identifies which rows (from which user) of
the whole given dataset should be extracted. This parameter is vectorised,
you may enter more \code{userid} at once.}
}
\value{
A list of numeric vectors with column names representing the attribute ID each
 reference point belongs to.
}
\description{
With Default Referenece Points we mean the initial configuration of the
product configurator, that is the value for each attribute that was selected
as default before the user interacted with the configurator. It assumes there
is a column with the name 'round' from which to read the default values (in
round = 0) for each user. The  The default values of each 'category' can be
interpreted as possible Reference Points for the decision maker. See Source
for more about this assumption.
}
\examples{
#Not Runnable yet
getDefaultRefps(camera2_config_data, c(6,9,10))
getDefaultRefps(as.data.frame(matrix_full), 55:58)
}
\seealso{
Other GetFunctions: \code{\link{benefitToCostAttr}},
  \code{\link{getAllUserIds}}, \code{\link{getAttrValues}},
  \code{\link{getRoundsById}}, \code{\link{getTableById}},
  \code{\link{get_attrs_ID}}, \code{\link{normalize}}
}

