% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_table_by_ID.R
\name{getget}
\alias{getget}
\title{Titel}
\usage{
getget(x, user_id)
}
\arguments{
\item{x}{the data set (table) you want to analize and separate in smaller data sets.}

\item{user_id}{an integer, that identifies which rows of the whole given dataset (x) should be extracted.}
}
\value{
A data frame with the same columns as the original, but only with rows about the provided \code{user_id}.
}
\description{
Description
}
\examples{
get_table_by_ID(camera2_config_data, 9)
}

