% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetFunctions.R
\name{getRoundsById}
\alias{getRoundsById}
\title{Get the amount of rounds (clicks) for each user}
\usage{
getRoundsById(dataset, userid = NULL)
}
\arguments{
\item{dataset}{the complete data.frame. You can also input the data.frame
specific for one user, such as the one generated with
\code{\link{getTableById}}.}

\item{userid}{an integer vector, that identifies which rows (from which user) of
the whole given dataset should be extracted. This parameter is vectorised,
you may enter more \code{userid} at once.}
}
\value{
A list of integer vectors representing the index of each round/click
 and therefore the total amount of rounds for the selected user.
}
\description{
Allows to calculate the amount of times a user interacted with the product
configurator, i.e. amount of clicks for any given \code{userid}. The entered
data.frame should signalize this information with a column named \code{round},
which should only contain numeric values. Function is vectorised in
\code{userid}.
}
\examples{
#Not Runnable yet
getRoundsById(example_data, userid = 6) # Example return: [1] 0 1 2 3 4 5 6 7 8
getRoundsById(as.data.frame(matrix_full), 12:10)

}
\seealso{
Other GetFunctions: \code{\link{benefitToCostAttr}},
  \code{\link{getAllUserIds}}, \code{\link{getAttrValues}},
  \code{\link{getDefaultRefps}},
  \code{\link{getTableById}}, \code{\link{get_attrs_ID}},
  \code{\link{normalize}}
}

