% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeightFunctions.R
\name{highestValue}
\alias{highestValue}
\title{Calculates weights relative to the highest sum relative to other attributes}
\usage{
highestValue(normalizedMatrix)
}
\arguments{
\item{normalizedMatrix}{}
}
\value{
a weight vector (numeric vector with a sum of 1)
}
\description{
Awards higher weights to those attirbutes which consistently showed higher
values, measured against the highest posssible. Secondary function used for
main function  \code{\link{weight.highAndStandard}}.
}
\examples{
#Runnable
highestValue(matrix(c(1.0, 0.85, 0.42, 0, 0.5, 0, 1, 0.7), 4, 2))
weights <- highestValue(matrix(c(1.0, 0.85, 0.42, 0, 0.5, 0, 1, 0.7), 2, 4))
sum(weights) should return 1
}

