% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PT-Functions.R
\name{lossFunction}
\alias{lossFunction}
\title{Calculates the loss for one single value}
\usage{
lossFunction(v1, v2)
}
\arguments{
\item{v1}{value of attribute j in round i}

\item{v2}{reference point of attribute j}
}
\value{
loss, numeric value.
}
\description{
Given the value of an attribute (v1) and the
reference point of the same attribute (v2), calculates the loss and returns it [1,2]. It
is not built as a stand alone function, rather as an object to be used by
other major functions, such as \code{gainMarix, lossMatrix}.
}
\details{
For understanding how this works, see the function itself or refer
to the paper. It handles only discrete numbers, so no interval numbers. Also
a point to improve further on.
}
\examples{
lossFunction(5, 1)  # returns: 0
lossFunction(2, 3)  # returns: -1

}
\references{
[1] Fan, Z. P., Zhang, X., Chen, F. D., & Liu, Y. (2013). Multiple
 attribute decision making considering aspiration-levels: A method based on
 prospect theory. Computers & Industrial Engineering, 65(2), 341-350.

 [2]Kahneman, D., & Tversky, A. (1979). Prospect theory: An analysis of
 decision under risk. Econometrica: Journal of the Econometric Society,
 263-291.
}

