% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PT-Functions.R
\name{lossMatrix}
\alias{lossMatrix}
\title{Loss matrix}
\usage{
lossMatrix(dataset, userid = NULL, attr = NULL, rounds = NULL,
  refps = NULL, cost_ids = NULL)
}
\arguments{
\item{userid}{an integer that gives the information of which user the matrix
should be calculated.}

\item{attr}{attributes IDs, vector of integer numbers corresponding to the
attributes you desire to use; attr are assumed to be 1-indexed.}

\item{rounds}{integer vector. Which steps of the configuration process should
be shown? See Details.}

\item{refps}{numeric vector. Reference Points: each point corresponds to one
attribute, i.e. each attribute has only one aspiration level. Default
setting assumes the aspiration levels as the default values of the initial
product configuration for each user.}

\item{cost_ids}{argument used to convert selected cost attributes into
benefit attributes. Integer vector.}

\item{data}{data.frame with the user generated data from a product
configurator. Please see \code{decision_matrix} for specifications of the
data.frame.}
}
\value{
a loss matrix for a specific user.
}
\description{
Creates the Loss matrix parting from a decision matrix and a vector
containing the reference points (aspiration-levels). We based our
calculations for gains and losses from a scientific paper, please see source
and references.
}
\details{
General: The loss_matrix with ncol = number of attributes you
selected or all(default) and nrow= number of rounds you selected or the first
and last(default) for a selected user.

\code{data} We assume the input data.frame has following columns usid = User
IDs, round = integers indicating which round the user is in (0-index works
best for 'round'), atid = integer column for referring the attribute ID (1
indexed), selected = numeric value of the attribute for a specific, given
round, selectable = amount of options the user can chose at a given round,
with the current configuration. This is a necessary parameter.

\code{userid} is a necessary parameter, without it you'll get a warning.
Default is NULL.

\code{attr} Default calculates with all attributes. Attributes are
automatically read from provided table, it is important you always provide
the complete dataset so that the package functions properly. Moreover the
attribute will be sorted in ascending order, i.e. if you input attr=
c(1,3,2), the decision matrix resulting will display the columns in order:
attr1 attr2 attr3.

\code{rounds} Default calculates first round(initia product config) and last
round of the session. Default calculates with first and last attributes
(initial and final product configuration). To choose all give "all" as
argument for rounds, see example. "first" and "last" are also possible
argument values. You can give a vector of arbitrarily chosen rounds as well.

\code{refps} If you only want to see the results for one attribute you may
enter only a couple of reference points but you have to tell the function
which attributes you want to use those referene points for. So the amount of
attr and of refps should be the same. Moreover the functions always orders de
attr, so be sure to input the reference point also in an ascending order
corresponding to their attributes. (refps will not be ordered)

\code{cost_ids} Default assumes all your attributes are of benefit type, that
is a higher value in the attribute means the user is better of than with a
lower value. If one or more of the attributes in your data is of cost type,
e.g. price, so that lower is better then you should identify this attributes
as such, providing their id, they'll be converted to benefit type (higher
amount is better).

This function is for one user only, for more or all users see
\code{\link{powerful_function}}
}
\examples{
loss_matrix(pc_config_data, 11)
loss_matrix(my_data, userid = 11, attr = c(1,3,5))
loss_matrix(keyboard_data, 60, rounds = "all", refps = c(1,3,4,0), cost_ids = 4)
loss_matrix(data1, 2, rounds = "last", attr = 1)
}

