% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestFunctions.R
\name{norm.gainLoss.sep}
\alias{norm.gainLoss.sep}
\title{Used for test-equal-normalized-matrix-function}
\usage{
norm.gainLoss.sep(dataset, userid = NULL, attr = NULL, rounds = NULL,
  refps = NULL, cost_ids = NULL)
}
\description{
Used for test-equal-normalized-matrix-function
}

