% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalizing.R
\name{normalize.altMethod}
\alias{normalize.altMethod}
\title{Alternative method for normalizing matrices}
\usage{
normalize.altMethod(aMatrix, attr, cost_ids = NULL)
}
\arguments{
\item{aMatrix}{a numeric (decision) matrix - if matrix has only one row, then
all values will be normalized to 1 even if the initial value is 0.}

\item{attr}{attributes IDs, vector of integer numbers corresponding to the
attributes you desire to use; attr are assumed to be 1-indexed. Even if not
all attributes are given the function normalizes all columns, by default as
benefit type unless shown otherwise with \code{cost_ids}.}

\item{cost_ids}{argument used to convert selected cost attributes into
benefit attributes. Integer vector.}
}
\value{
a normalized matrix
}
\description{
This function is used as the first step in
\code{\link{weight.differenceToIdeal}} and \code{\link{weight.entropy}} for normalizing a matrix.
}
\details{
This functions uses an alternative approach to the one used in
  \code{\link{gainLoss}} in regard to \code{cost_ids}. Instead of multiplying all
  column values with (-1) so that lower values become the higher values, it
  normalizes relative to the lowest value, as given by [1] and [2]. Main
  difference in the result is that this normalizing function does not return
  negative values.
}
\examples{
# Runnable
normalize.altMethod(matrix(1:16, 4, 4))
normalize.altMethod(matrix(1:16, 4, 4), attr=1:4, cost_ids = c(2,4))

}
\references{
[1]Ma, J., Fan, Z. P., & Huang, L. H. (1999). A subjective and
  objective integrated approach to determine attribute weights. European
  journal of operational research, 112(2), 397-404.

  [2] Fan, Z. P. (1996). Complicated multiple attribute decision making:
  theory and applications (Doctoral dissertation, Ph. D. Dissertation,
  North-eastern University, Shenyang, PRC).
}

