% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalizing.R
\name{normalize.highestValue}
\alias{normalize.highestValue}
\title{Auxiliary function for calculating weights}
\usage{
normalize.highestValue(aMatrix, attr, cost_ids = NULL)
}
\arguments{
\item{aMatrix}{a matrix to be 'normalized'}

\item{attr}{attributes IDs, vector of integer numbers corresponding to the
attributes you desire to use; attr are assumed to be 1-indexed. Even if not
all attributes are given the function normalizes all columns, by default as
benefit type unless shown otherwise with \code{cost_ids}.}

\item{cost_ids}{argument used to convert selected cost attributes into
benefit attributes. Integer vector.}
}
\value{
a normalized matrix
}
\description{
Does not need full documentation, since not primarly intended for users, but
as an auxiliary function to \code{\link{weight.highestValue}}.
}

