% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalizing.R
\name{normalize.sum}
\alias{normalize.sum}
\title{Title}
\usage{
normalize.sum(aMatrix)
}
\arguments{
\item{aMatrix}{}

\item{aMatrix}{a numeric (decision) matrix - if matrix has only one row, then
all values will be normalized to 1 even if the initial value is 0.}
}
\value{


a normalized matrix
}
\description{
This function is indifferent as to where it is a cost_ids or not, since the
same formula works for both. It does have a limitation, i.e. if on a single
attribute, there are simultaneously negative and positive values, in details?
}
\details{
Note: this function has one important limitation, if within a same
  attribute there are negative and positive values, the function will likely
  produce a \code{NaN}.

This functions uses the same equation to normalize benefit and cost
  type attributes. It returns values between 0 and 1, with the only
  limitation that when an attribute has both negative and positive values, it
  may return a \code{NaN}.
}
\examples{
Normalizes a function according to the entropy method

This function helps realize the first step in \\code{\\link{weight.entropy}}
within the entropy method for calculating weights, as shown for example by
[1] and [2].

}
\references{
[1]Lotfi, F. H., and Fallahnejad, R. (2010). Imprecise Shannons entropy and
  multi attribute decision making. Entropy, 12(1), 53-62.

  [2] Fan, Z. P. (1996). Complicated multiple attribute decision making: theory
  and applications (Doctoral dissertation, Ph. D. Dissertation, North-eastern
  University, Shenyang, PRC).

  [2] Hwang, C. L., & Yoon, K. (2012). Multiple attribute decision making:
  methods and applications a state-of-the-art survey (Vol. 186). Springer
  Science & Business Media.
}

