% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prospect_value_functions.R
\name{overall_pv}
\alias{overall_pv}
\title{Takes}
\usage{
overall_pv(dataset, userid = NULL, attr = NULL, rounds = NULL,
  refps = NULL, cost_ids = NULL, weight = NULL, alpha = 0.88,
  beta = 0.88, lambda = 2.25)
}
\arguments{
\item{userid}{an integer that gives the information of which user you want the data from: User ID.}

\item{attr}{attributes IDs, vector of integer numbers corresponding to the attributes you desire to use;1-indexed.}

\item{rounds}{integer vector. Which steps of the configuration process should be shown? See Details.}

\item{refps}{numeric vector. Reference Points: each point corresponds to one attribute, i.e. each attribute has only one
aspiration level.}

\item{cost_ids}{argument used to convert selected cost attributes into benefit attributes. Integer vector.}

\item{weight}{orders each attribute a weight <= 1 for the value functions to assign a value to each alternative or in the case of productConfig, to each round.}

\item{alpha}{numeric between [0, 1]. Determines the concativity of the value function as given by Reference[1].}

\item{beta}{numeric between [0, 1]. Determines the convexity of the value function as given by Reference[1].}

\item{lambda}{lambda > 1. Parameter of loss aversion for the value function as given by Reference[1].}

\item{data}{data.frame with the user generated data from a product configurator. See Details
 for the specifications of the data.frame.}
}
\value{
The same value that returns the function you entered in \code{FUN} for the users you entered in the \code{userid} parameter.
}
\description{
Pro
}
\details{
If you need more detailed information about the parameters please see \code{\link{pvalue_matrix}}

\code{data} We assume the input data.frame has following columns usid = User IDs, round = integers indicating which round the user is in
(0-index works best for 'round'), atid = integer column for referring the attribute ID (1 indexed), selected = numeric value of the attribute for a specific, given round,
selectable = amount of options the user can chose at a given round, with the current configuration. This is a necessary parameter.

\code{userid} is a necessary parameter, without it you'll get a warning. Default is NULL.

\code{weight}

\code{alpha} Default value as given by Reference [1] is 0.88

\code{beta} Default value as given by Reference [1] is 0.88

\code{lambda} Default value as given by Reference [1] is 2.25
}
\examples{
powerful_function(data_camera, 2)  # FUN=decision_matrix
powerful_function(data_pc, userid = c(1,2,4,8,9,100))
powerful_function(my_table, get_all_userids(my_table))

powerful_function(data_pc, userid = c(1,2,3), FUN = gain_matrix)
powerful_function(data_pc, userid = c(1,2,3), FUN = overall_pv, alpha = 0.9, weight=c(0.1,0.2,0.5,0.2))
powerful_function(datadata, c(6,9), FUN = ref_points) # Returns the default values for both users (user6, user9) assumed as reference points.
}
\seealso{
\code{\link{powerful_function}}
}

