% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProspectValueFunctions.R
\name{overall_pv_extend}
\alias{overall_pv_extend}
\title{Calculate Prospect Values}
\usage{
overall_pv_extend(value_matrix, weight = NULL)
}
\arguments{
\item{value_matrix}{a List of numeric matrices, results from using the value function on
previously calculated normalized gain and loss matrices.}

\item{weight}{numeric vector. Represents the importance or relevance that an
attribute has and the weight it should have in the calculation of the
prospect value. Alternatively, you can enter a list of numeric vectors, each
element of the list corresponding to one user in \code{userid}.}
}
\value{
a list of overall prospect values for each round (each product
  alternative). Each element of a list is the result for one user.
}
\description{
This function works with the simple additive weighting method. It takes a
value matrix and the weight of each attribute and calculates the prospect
value for each alternative (round) [1,2].
}
\details{
You need to pre-calculate the value matrix, for example with
  \code{\link{pvalue_matrix}} and give it as a parameter. This is one of the
  few functions of this package that do not allow you to give the raw data
  from your product Configurator, but rather calculate a previous
  result(value matrix) to input here.

  \code{value_matrix} ncol = number of attributes, nrow = number of rounds.
}
\examples{
overall_pv_extend(my_value_matrix, weight = c(0.1, 0.2, 0.4, 0.3))
overall_pv_extend(pvMatrix(someData, someUsers), weight = getAttrWeights(someData, someUsers))
overall_pv_extend(my_matrix, weight=c(0.8,0.05,0.05,0.1))
overall_pv_extend(vMatrix, weight =  aWeightVector)

\dontrun{overall_pv_extend(value_mx)}  # Always provide weights or dataset.

}
\references{
[1] Fan, Z. P., Zhang, X., Chen, F. D., & Liu, Y. (2013).
  Multiple attribute decision making considering aspiration-levels: A method
  based on prospect theory. Computers & Industrial Engineering, 65(2),
  341-350.

  [2] Kahneman, D., & Tversky, A. (1979). Prospect theory: An analysis of
  decision under risk. Econometrica: Journal of the Econometric Society,
  263-291.
}
\seealso{
\code{\link{overallPV}}
}

