% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProspectValueFunctions.R
\name{overallPV}
\alias{overallPV}
\title{Calcultes the overall prospect values using PT (prospect theory)}
\usage{
overallPV(dataset, userid = NULL, attr = NULL, rounds = NULL,
  refps = NULL, cost_ids = NULL, weight = NULL,
  weightFUN = "differenceToIdeal", alpha = 0.88, beta = 0.88,
  lambda = 2.25, gamma)
}
\arguments{
\item{dataset}{a \code{data.frame} with the user generated data from a
product configurator. See Details of \code{\link{decisionMatrix}} for more
information about which data should included in this argument.}

\item{userid}{an integer vector indicating for which user the output of this
function should be calculated. This functions is vectorised in this
argument, i.e. you may enter more userIDs simultaneously.}

\item{attr}{attribute IDs, vector of integer numbers corresponding to the
attributes (columns) you desire to use.}

\item{rounds}{integer vector, text option or a list of integer vectors. Which
steps of the configuration process should be shown? Defaults are first and
last step. Text options are \code{all, first, last}. Alternatively, a
vector of arbitrarily chosen rounds can be entered as well.}

\item{refps}{a list of numeric vectors, one for each user. Reference Points:
each point corresponds to one attribute, therefore the amount of attributes
and of refps entered, should be equal. Default assumes the refps as the
default values of the initial product configuration for each user. You may
fully or partially enter your own reference points, check below for more
info.}

\item{cost_ids}{argument used to convert selected cost attributes into
benefit attributes. Integer vector. Cost type attributes have the
characteristic, that a lower value means the user is better off than with a
higher value. E.g. price is often considered a cost type attribute. Should be equal to
\code{attr} input or a subset of it.}

\item{weight}{numeric vector. Represents the importance or relevance that an
attribute has and the weight it should have in the calculation of the
prospect value. Alternatively, you can enter a list of numeric vectors, each
element of the list corresponding to one user in \code{userid}.}

\item{weightFUN}{indicated which weight function should be used to calculate
the weight vector, the options are \code{"differenceToIdeal", "entropy",
"highAndStandard", "standard"}. Default ist \code{differencetoIdeal}.}

\item{alpha}{numeric between [0, 1]. Determines the concativity of the value
function and has a default value of 0.88 as given by Reference[1].}

\item{beta}{numeric between [0, 1]. Determines the convexity of the value
function and has a default value of 0.88  as given by Reference[1].}

\item{lambda}{lambda > 1. Parameter of loss aversion for the value function
as and has a default value of 2.25 given by Reference[1].}

\item{gamma}{numeric and between 0 and 1. It is a parameter used for the function
\code{\link{weight.highAndStandard}} and
\code{\link{highAndStandard}}, please the former to know its use.}
}
\value{
overall prospect values for each attribute
}
\description{
(1) It calculates the normalized gain and loss matrices, (2) with both
matrices the value matrix is then calculated and finally (3) the prospect
value for each alternative/round/row. From step (2) to (3) this function uses
prospect theory's (PT) value function[1,2] and a method proposed by [3]. Its
equivalent function for multi-reference point approaches are
\code{\link{overallDRP}} and \code{\link{overallTRP}}.
}
\details{
In the context of data stemming from product configurators, the highest
overall value returned  by this function, means that specific product
configuration represented the highest value for the user. This depends mainly
on three important factors: (1) What theoretical framework you choose to use
(PT, DRP, TRP), (2) which decision weights you assign to each attribute, and
(3) the reference point(s) you input.

This function is vectorized in the \code{userid} parameter. The
  function does not sort attributes or user IDs. Order of the output is
  generated as given through the arguments.

  The 3 step calculation of the prospect values comes from one specific paper
  \emph{Reference[1]}. (1) For the noramlized gain and loss matrices we use
  the function \code{\link{norm.gainLoss}} from this package. (2) The value
  matrix is calculated with a series of auxiliary functions. (3) The prospect
  value works with a simple additive weighting method from
  \code{overall_pv_extend}.

  \code{dataset} We assume the input data.frame has following columns usid =
  User IDs, round = integers indicating which round the user is in (0-index
  works best for 'round'), atid = integer column for referring the attribute
  ID (1 indexed), selected = numeric value of the attribute for a specific,
  given round, selectable = amount of options the user can chose at a given
  round, with the current configuration. This is a necessary parameter.

  \code{userid} is a necessary parameter.

  \code{weight} default orders each attribute a weight <= 1 according to the
  the weight function \code{\link{differenceToIdeal}}. Ideally the sum of all
  weights equals 1.

  \code{alpha} Default value as given by Reference [1] is 0.88

  \code{beta} Default value as given by Reference [1] is 0.88

  \code{lambda} Default value as given by Reference [1] is 2.25
}
\examples{
#Not runnable yet
overallPV(data_camera, 2)
overallPV(data_pc, userid = 1, weight=c(0.1,0.4,0.3,0.2))
overallPV(full_data, 6 ,attr = c(1,2,3), rounds="all", alpha = 0.95, beta = 0.78)

}
\references{
[1]Kahneman, D., & Tversky, A. (1979). Prospect theory: An
  analysis of decision under risk. Econometrica: Journal of the Econometric
  Society, 263-291.

  [2] Tversky, A., & Kahneman, D. (1992). Advances in prospect theory:
  Cumulative representation of uncertainty. Journal of Risk and uncertainty,
  5(4), 297-323.

  [3] Fan, Z. P., Zhang, X., Chen, F. D., & Liu, Y. (2013). Multiple
  attribute decision making considering aspiration-levels: A method based on
  prospect theory. Computers & Industrial Engineering, 65(2), 341-350.
}
\seealso{
\code{\link{decisionMatrix, overallTRP, overallDRP,
  weight.differenceToIdeal, weight.entropy, weight.highAndStandard}}
}

