% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProspectValueFunctions.R
\name{overallTRP}
\alias{overallTRP}
\title{Calcultes the overall prospect values using the TRP approach (tri reference
point theory)}
\usage{
overallTRP(dataset, userid = NULL, attr = NULL, rounds = NULL,
  refps = NULL, cost_ids = NULL, weight = NULL,
  weightFUN = "differenceToIdeal", tri.refps = NULL, beta_f = 5,
  beta_l = 1, beta_g = 1, beta_s = 3, gamma)
}
\arguments{
\item{dataset}{data.frame with the user generated data from a product
configurator. See \code{decisionMatrix} for specifications of the dataset.}

\item{userid}{a vector of integers that gives the information of which users
the matrix should be calculated. Vectorised.}

\item{attr}{attributes IDs, vector of integer numbers corresponding to the
attributes you desire to use; attr are assumed to be 1-indexed.}

\item{rounds}{integer vector or text option. Which steps of the configuration
process should be shown? Defaults are first and last step. Text options are
\code{all, first, last}.}

\item{cost_ids}{argument used to convert selected cost attributes into
benefit attributes. Integer vector.}

\item{weight}{numeric vector. Represents the importance or relevance that an
attribute has and the weight it should have in the calculation of the
prospect value.}

\item{weightFUN}{indicated which weight function should be used to calculate
the weight vector, the options are \code{"differenceToIdeal", "entropy",
"highAndStandard", "standard"}. Default ist \code{differencetoIdeal}.}

\item{gamma}{numeric and between 0 and 1. It is a parameter used for the function
\code{\link{weight.highAndStandard}} and
\code{\link{highAndStandard}}, please the former to know its use.}
}
\value{
overall prospect values for each alternative/row
}
\description{
The overall prospect values for each alternative (round) are calculated using
two components in a method described in [2]: (1) the value matrix, which is
calculated using \code{\link{trp.valueMatrix}} with the trp-value function as
given by [1]. And the (2) decision weights, which can be calculated in this
package using three functions. See Details of weight attribute. Its
equivalent function for prospect theory and the dual reference point approach
are \code{\link{overallPV}} and \code{\link{overallDRP}}.
}
\examples{
#Not runnable yet
overallTRP(pc_config_data, 9, attr=3,  tri.refps = c(1, 3.5))
overallTRP(pc_config_data, 9, attr=1 tri.refps = c(1, 3.5), lambda=2, delta=0.5)
overallTRP(myData, userid = 58, rounds = "all", attr=3:4, tri.refps=matrix(1:6, 2, 3, byrow=T))
overallTRP(myData, userid = 11, attr =  1, cost_ids = 1, tri.refps = c(8, 2))
overallTRP(full_data, 30:35 ,attr = c(1,2,3), rounds="all", tri.refps=matrix(1:9, 3, 3, byrow=T))

}
\references{
[1] [1]Wang, X. T.; Johnson, Joseph G. (2012) \emph{A tri-reference
  point theory of decision making under risk. }Journal of Experimental
  Psychology

  [2] Fan, Z. P., Zhang, X., Chen, F. D., & Liu, Y. (2013). Multiple
  attribute decision making considering aspiration-levels: A method based on
  prospect theory. Computers & Industrial Engineering, 65(2), 341-350.
}
\seealso{
\code{\link{decisionMatrix, overallTRP, overallDRP,
  weight.differenceToIdeal, weight.entropy, weight.highAndStandard}}
}

