% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PT-Functions.R
\name{pvalue_fun}
\alias{pvalue_fun}
\title{Implements prospect theory's value function}
\usage{
pvalue_fun(ngain_ij, nloss_ij, alpha = 0.88, beta = 0.88, lambda = 2.25)
}
\arguments{
\item{ngain_ij}{gain value corresponding to a specific attribute (j) and round (i)}

\item{nloss_ij}{loss value corresponding to the same specific attribute and round as \code{ngain_ij}}

\item{alpha}{parameter for diminishing sensitivity in the gain domain. Default value = 0.88. Usual values are in the (0,1) interval.}

\item{beta}{parameter for diminishing sensitivity in the loss domain. Default value = 0.88. Usual values are in the (0,1) interval.}

\item{lambda}{parameter for loss aversion. Default value = 2.25. Values for \code{lambda} should be > 1.}
}
\value{
the output of the value function [1]
}
\description{
Given gains and losses relative to a reference points, we use prospect
theory's value function to calculate the value matrix[1]. We use default values for the
parameters of diminishing sensitity and loss aversion, but this can be
inputed differently. Calculates the value function for one value at a time. The function
is used to calculate the value matrix [2] in \code{\link{pvMatrix}}.
}
\examples{
pvalue_fun(3, 0)
pvalue_fun(0, -1.5, alpha =0.75, lambda=3)

}
\references{
[1] Kahneman, D., & Tversky, A. (1979). Prospect theory: An
  analysis of decision under risk. Econometrica: Journal of the Econometric
  Society, 263-291.

  [2] Fan, Z. P., Zhang, X., Chen, F. D., & Liu, Y. (2013). Multiple
  attribute decision making considering aspiration-levels: A method based on
  prospect theory. Computers & Industrial Engineering, 65(2), 341-350.
}

