% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProspectValueFunctions.R
\name{pvalue_matrix}
\alias{pvalue_matrix}
\title{Calcultes the Value Matrix}
\usage{
pvalue_matrix(dataset, userid = NULL, attr = NULL, rounds = NULL,
  refps = NULL, cost_ids = NULL, alpha = 0.88, beta = 0.88,
  lambda = 2.25)
}
\arguments{
\item{dataset}{data.frame with the user generated data from a product
configurator. See Details for the specifications of the data.frame.}

\item{userid}{an integer that gives the information of which user you want
the data from: User ID.}

\item{attr}{attributes IDs, vector of integer numbers corresponding to the
attributes you desire to use;1-indexed.}

\item{rounds}{integer vector. Which steps of the configuration process should
be shown? See Details.}

\item{refps}{numeric vector. Reference Points: each point corresponds to one
attribute, i.e. each attribute has only one aspiration level.}

\item{cost_ids}{argument used to convert selected cost attributes into
benefit attributes. Integer vector.}

\item{alpha}{numeric between [0, 1]. Determines the concativity of the value
function as given by Reference[1].}

\item{beta}{numeric between [0, 1]. Determines the convexity of the value
function as given by Reference[1].}

\item{lambda}{lambda > 1. Parameter of loss aversion for the value function
as given by Reference[1].}
}
\value{
the value matrix
}
\description{
According to the parameters, it first calculates the normalized gain and loss
matrices. Using \code{\link{prospect_value_matrix_extend}} it calculates the
value matrix using the value function given by Tversky & Kahnemann(1992)[1].
}
\details{
This function is for one user or one userid for \strong{more users}
and for more detailed \strong{parameter information} please see
\code{\link{powerful_function}}

\code{dataset} We assume the input data.frame has following columns usid =
User IDs, round = integers indicating which round the user is in (0-index
works best for 'round'), atid = integer column for referring the attribute ID
(1 indexed), selected = numeric value of the attribute for a specific, given
round, selectable = amount of options the user can chose at a given round,
with the current configuration. This is a necessary parameter.

\code{userid} is a necessary parameter.
}
\examples{
pvalue_matrix(data_camera, 2,)
pvalue_matrix(data_pc, 100, weight=c(0.1,0.4,0.3,0.2))
pvalue_matrix(full_data, userid = 25 ,alpha = 0.95, beta = 0.78)

}
\seealso{
\code{\link{powerful_function}}

Other prospect value functions: \code{\link{overall_pv_extend}},
  \code{\link{overall_pv}},
  \code{\link{prospect_value_matrix_extend}}
}

