% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetFunctions.R
\name{replaceNotNA}
\alias{replaceNotNA}
\title{Auxiliary function used for testing.}
\usage{
replaceNotNA(x, y, boolean.vector)
}
\arguments{
\item{x}{an element with numeric, logical or char values.}

\item{y}{an element with numeric, logical or char values.}

\item{boolean.vector}{containing indexes of which values should be changed.}
}
\value{
a subset of \code{x}
}
\description{
Auxiliary function used for testing.
}
\examples{
replaceNotNA(a, b, c(T, F, T, T))
}

