% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRPFunctions.R
\name{smallerThanZero}
\alias{smallerThanZero}
\title{Outputs a value matrix from a decision matrix}
\usage{
smallerThanZero(aMatrix, dual.refps, lambda = 2.25, delta = 0.8)
}
\arguments{
\item{aMatrix}{the decision matrix, with \code{attr} as columns and
\code{rounds} as rows.}

\item{dual.refps}{two numeric reference points (status-quo, goal).}

\item{lambda}{numeric - parameter of loss aversion for the value function as
given by reference[1]. Default value is 2.25 as given by [2].}

\item{delta}{numeric - expresses the relative importance of the aspiration
level to other factors. Default is 0.8 and it should satisfy \code{0 <
delta <1}.}
}
\value{
a value matrix with equal dimensions as the input \code{aMatrix}
}
\description{
This function is called within \code{\link{trpValueMatrix}}, but it is the
function that actually does the work of producing the value matrix.It works
in three steps. The first step checks if the matrix or the reference points
have a value smaller or equal to 0. If this is the case it monotonically
scales all values so that they are all larger than zero*. The second step
calculates a gain-loss matrix using both reference points and a function
given by [1]. The third step normalizes the gain-loss matrix to make all
attributes comparable.
}
\details{
*The transformation changes all values, but keeps the differences
  between them as they were.
}
\examples{
#Runnable
smallerThanZero(matrix(16:31, 4, 4), dual.refps= c(22, 28))
smallerThanZero(matrix(16:31, 4, 4), dual.refps= c(22, 28), delta= 0.6)
smallerThanZero(matrix(1:100, 5, 20, byrow= T), dual.refps= c(sq=45, g=88))

#Not runnable yet
smallerThanZero <- function(decisionMatrix(myData, 9, rounds="all"), c(1.5, 2.5), lambda = 5, delta = 0.8)
}
\references{
[1] Golman, R., & Loewenstein, G. (2011). Explaining Nonconvex
  Preferences with Aspirational and Status Quo Reference Dependence. Mimeo,
  Carnegie Mellon University.

  [2] Tversky, A., & Kahneman, D. (1992). Advances in prospect theory:
  Cumulative representation of uncertainty. Journal of Risk and uncertainty,
  5(4), 297-323.
}

