% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeightFunctions.R
\name{standardDeviation}
\alias{standardDeviation}
\title{Calculates decision weights using the standard deviation method}
\usage{
standardDeviation(aMatrix)
}
\arguments{
\item{normalizedMatrix}{a numeric, normalized matrix. If indeed normalized it should only
contain values between \code{0} and \code{1}.}
}
\value{
a weight vector(numeric vector with a sum of 1)
}
\description{
This function is used as the second step in \code{\link{weight.standard}} for
calculating a decision weight for each \code{\link{attr}} in the decision
matrix. The methodology of this method for determining the weights
out of a decision matrix is given by references [1] and [2]. See References.
}
\details{
The sum of the output of this functions should always equal 1.

Similar to \code{\link{entropy}} small differences
  between value attributes are rewarded a lower value and thus a relative
  lower weight.
}
\examples{
#Runnable
standardDeviation(matrix(c(1.0, 0.85, 0.42, 0, 0.5, 0, 1, 0.7), 4, 2))
weights <- standardDeviation(matrix(c(1.0, 0.85, 0.42, 0, 0.5, 0, 1, 0.7), 2, 4))
sum(weights) should return 1

}
\references{
[1]Diakoulaki, D., Mavrotas, G., & Papayannakis, L. (1995).
  Determining objective weights in multiple criteria problems: the CRITIC
  method. Computers & Operations Research, 22(7), 763-770.

  [2]Jahan, A., & Edwards, K. L. (2013). Multi-criteria decision analysis for
  supporting the selection of engineering materials in product design.
  Butterworth-Heinemann.
}

