% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRPFunctions.R
\name{trp.overallPV_extend}
\alias{trp.overallPV_extend}
\title{cost_ids, enter normal reference points, we will convert them. cost_ids has to equal the attribute you are inputting
DOCU: Converting tri.refps, no need to convert if attribute is of type cost.
New function as interface with weights and your trp.valueMatrix
Docu: For the way the trp function works it is a little more complicated than for overallPV for the pt
here we have to manually calculate the AttributeWeights whit your desired function, e.g ww <- getAttrWeights(...)
and the trp.ValueMatrix separately as well, trp.VM <- mapply() OR trpValueMatrix(...)
then giving to this function as input and getting the desired result}
\usage{
trp.overallPV_extend(trp.ValueMatrix, weight = NULL)
}
\description{
DOCU: Explain what _extends is in pC, singalizes major functions that do not take the normal inputs but user
other functions' results to work.
}

