% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRPFunctions.R
\name{trpValueFunction}
\alias{trpValueFunction}
\title{Transform a decision Matrix into a trp value matrix}
\usage{
trpValueFunction(aMatrix, triRefps, beta_f = 5, beta_l = 1.5, beta_g = 1,
  beta_s = 3)
}
\arguments{
\item{aMatrix}{a non-empty matrix, tipically with one column since this
function is called one attribute at a time by
\code{trpValueMatrix.oneAttr}.}

\item{mr}{numeric - Minimum Requirements is the lowest reference point}

\item{sq}{numeric - Status Quo reference point}

\item{g}{numeric - Goal reference point}

\item{beta(s)}{numeric arguments representing the psychological impact of an
outcome equaling failer (_f), loss (_l), gain (_g) or success (_s). Default
values are taken from our reference paper \code{(5,1,1,3)}. See details.}
}
\value{
returns a matrix with the outputs of the trp value function for each of its elements
}
\description{
This function is based on the value function of the tri-reference point (trp)
theory. It is an auxiliary function, which intends to facilitate the work and
readability of \code{\link{trpValueMatrix.oneAttr, trpValueMatrix}}. It takes
a matrix and the three given reference points (MR, SQ, G) as a vector
\code{tri.refps} and applys the trp value function \code{trpValueFunction} to
each element of the matrix. Also takes into account the for free \code{beta}
parameters of the function.
}
\details{
The functions test for MR < SQ < G

The beta arguments are important arguments that give form to the value function proposed in [1].
A higher number represents a higher relative psychological impact to the decision maker. Since in [1] it is assumed that the
reference point 'Minimum Requierment' has a greater impact when is not reached (failure aversion), it should have a higher beta, so in general
\code{beta_f > beta_l > beta_g > beta_s}. See our reference paper for a detailed theoretical background.
}
\examples{
# Runnable
trpValueFunction(aMatrix = matrix(1:6, 2, 3), triRefps = c(2,3,4.5))
trpValueFunction(matrix(1:16, 16, 1), triRefps = c(4, 8.9, 12.5), beta_f = 7)

}
\references{
[1] Wang, X. T.; Johnson, Joseph G. (2012) \emph{A tri-reference
  point theory of decision making under risk. }Journal of Experimental
  Psychology

  [2]Wang, X. T.; Johnson, Joseph G. (2012) \emph{Supplemental Material for: A tri-reference
  point theory of decision making under risk. }Journal of Experimental
  Psychology
}

