% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRPFunctions.R
\name{trpValueFunction_extend}
\alias{trpValueFunction_extend}
\title{TTri Reference Point Value Function for one element}
\usage{
trpValueFunction_extend(x, mr = 0.5, sq = 1.5, g = 2.5, beta_f = 5,
  beta_l = 1, beta_g = 1, beta_s = 3)
}
\arguments{
\item{x}{one numeric value}

\item{mr}{numeric - Minimum Requirements is the lowest reference point}

\item{sq}{numeric - Status Quo reference point}

\item{g}{numeric - Goal reference point}

\item{beta(s)}{numeric arguments representing the psychological impact of an
outcome equaling failer (_f), loss (_l), gain (_g) or success (_s). Default
values are taken from our reference paper \code{(5,1,1,3)}. See references.}
}
\value{
the output of v(x) with v: trp value function([1]).
}
\description{
Auxiliary function: it is based on the value function of the tri-reference
point (trp) theory. It's called by \code{\link{trpValueFunction}}, it takes
one element and puts it through the trp value function as seen in reference
[1]. Not vectorised.
}
\details{
The functions test for MR < SQ < G.

  The beta arguments are important arguments that give form to the value
  function proposed in [1]. A higher number represents a higher relative
  psychological impact to the decision maker. Since in [1] it is assumed that
  the reference point 'Minimum Requierment' has a greater impact when is not
  reached (failure aversion), it should have a higher beta, so in general
  \code{beta_f > beta_l > beta_g > beta_s}. See our reference paper for a
  detailed theoretical background.

  On reference points by cost type \code{attr}: For a cost attribute it should be
  true, that a lower value is better for the user, this should also hold for
  the three reference points. So contrary to normal/benefit attributes \code{
  for cost attributes} reference points should follow that: \code{mr > sq >
  g}.
}
\examples{
# Runnable
trpValueFunction_extend(0.18, mr = 0.15, sq = 0.55, g = 1.10)
trpValueFunction_extend(4, mr = 1, sq = 3, g = 8, beta_f = 7, beta_s = 4)

}
\references{
[1] Wang, X. T.; Johnson, Joseph G. (2012) \emph{A tri-reference
  point theory of decision making under risk. }Journal of Experimental
  Psychology

  [2]Wang, X. T.; Johnson, Joseph G. (2012) \emph{Supplemental Material for: A tri-reference
  point theory of decision making under risk. }Journal of Experimental
  Psychology
}

