% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRPFunctions.R
\name{trpValueMatrix}
\alias{trpValueMatrix}
\title{Returns a Value Matrix using three reference points}
\usage{
trpValueMatrix(dataset, userid = NULL, attr = NULL, rounds = NULL,
  cost_ids = NULL, mr = 0.5, sq = 1.5, g = 2.5, beta_f = 5,
  beta_l = 1, beta_g = 1, beta_s = 3)
}
\arguments{
\item{dataset}{data.frame with the user generated data from a product
configurator. See \code{decision_matrix} for specifications of the dataset.}

\item{userid}{a vector of integers that gives the information of which users
the matrix should be calculated. Vectorised.}

\item{attr}{attributes IDs, vector of integer numbers corresponding to the
attributes you desire to use; attr are assumed to be 1-indexed.}

\item{rounds}{integer vector or text option. Which steps of the configuration
process should be shown? Defaults are first and last step. Text options are
\code{all, first, last}.}

\item{cost_ids}{argument used to convert selected cost attributes into
benefit attributes. Integer vector.}

\item{mr}{numeric - Minimum Requirements is the lowest reference point}

\item{sq}{numeric - Status Quo reference point}

\item{g}{numeric - Goal reference point}

\item{beta(s)}{numeric arguments representing the psychological impact of an outcome
equaling failer (_f), loss (_l), gain (_g) or success (_s). Default values
are taken from our reference paper \code{(5,1,1,3)}.}
}
\value{
a list of value matrices for each user.
}
\description{
This function is based on the value function of the tri-reference point (trp)
theory. It first builds a desicion matrix for each user and then applys the
trp-value function over each value using the three given reference points
(MR, SQ, G) and other four free parameters from the value function. See
references.
}
\details{
This function only makes sense to use with multiple attributes, it
  those attributes have exactly the same three reference points (mr, sq, g).
  Therefore you will have to manually calculate all the value matrices for
  the different attributes (with different values) and cbind them together
  using mapply. The full matrix can then be given as an input to the
  \code{\link{overallPV_interface}} fucntion to calculate the overall
  prospect values for each round.

  General: The value matrix has ncol = number of attributes you selected or
  all(default) and nrow = number of rounds you selected or the first and
  last(default) for all selected users.

  \code{data} We assume the input data.frame has following columns usid =
  User IDs, round = integers indicating which round the user is in (0-index
  works best for 'round'), atid = integer column for referring the attribute
  ID (1 indexed), selected = numeric value of the attribute for a specific,
  given round, selectable = amount of options the user can chose at a given
  round, with the current configuration.

  \code{userid} is a necessary parameter, without it you'll get a warning.
  Default is NULL.

  \code{attr} Default calculates with all attributes. Attributes are
  automatically read from provided dataset, it is important you always
  provide the complete data so that the package functions properly. Moreover,
  \code{userid} and \code{attr} will not be sorted and will appear in the
  order you input them.

  \code{rounds} Default calculates with first and last rounds (initial and
  final product configuration). You can give a vector of arbitrarily chosen
  rounds as well.

  \code{cost_ids} Default assumes all your attributes are of benefit type,
  that is a higher value in the attribute means the user is better off than
  with a lower value. If one or more of the attributes in your data is of
  cost type, e.g. price, so that lower is better then you should identify
  this attributes as such, providing their id, they'll be converted to
  benefit type (higher amount is better).

  Note: When converting a cost attribute to a benefit attribute its three reference points change as well, enter the
  unconverted refps, the function transforms them automatically when it detects a \code{cost_ids}
}
\examples{
trpValueMatrix(pc_config_data, 9:11, mr = 0.5, sq = 0, g = 2.5)
trpValueMatrix(aDataFrame, userid = 100, rounds = "all", mr = 0.5, sq = 0, g = 2.5)
trpValueMatrix(my_data, userid = 11, attr = c(1,3,5), cost_ids = 1:3) #Input accepted but cost_ids = 2 will be ignored
trpValueMatrix(keyboard_data, 60, rounds = "first", attr=1, mr = 0.5, sq = 0, g = 2.5, beta_f = 6)
gain_matrix(data1, 2) # Returns an error since no reference points entered (mr, sq, g)

}
\references{
Wang, X. T.; Johnson, Joseph G. (2012) \emph{A tri-reference
  point theory of decision making under risk. }Journal of Experimental
  Psychology
}

