% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeightFunctions.R
\name{weight.highAndStandard}
\alias{weight.highAndStandard}
\title{Calculates weights using two weighted sub-functions}
\usage{
weight.highAndStandard(dataset, userid = NULL, attr = NULL,
  rounds = "all", cost_ids = NULL, gamma = 0.5)
}
\arguments{
\item{dataset}{data.frame with the user generated data from a product
configurator. See \code{decisionMatrix} for specifications of the dataset.}

\item{userid}{a vector of integers that gives the information of which users
the matrix should be calculated. Vectorised.}

\item{attr}{attributes IDs, vector of integer numbers corresponding to the
  attributes you desire to use; attr are assumed to be 1-indexed. This
  function will calculate with all attributes and do the subsetting a
  posteriori.

  If you want to get the weights for only two attributes you will have to
  first use \code{\link{decisionMatrix}} and then pass it on to
  \code{\link{normalize.altMethod}} and \code{\link{differenceToIdeal}}.}

\item{rounds}{integer vector or text option. Which steps of the configuration
process should be taken into account? Defaults are "all" in order to have
more data to calculate with. If \code{"first"} or \code{"last"} are entered
there will be only one rounds to gather data from, consequently all
attribtues will have the same weight.}

\item{cost_ids}{argument used to convert selected cost attributes into benefit
attributes. Integer vector.}

\item{gamma}{numeric and between 0 and 1. It is a parameter used for the function
\code{\link{weight.highAndStandard}} and
\code{\link{highAndStandard}}, please the former to know its use.}
}
\value{
a list of weight vectors (one per user)
}
\description{
This function calculates two separate weight vectors and merges them together
with a weighted parameter \code{gamma}.
}
\details{
The first weight function \code{\link{weight.highestValue}} rewards those
attributes which have values (consistenly) closer to the highest possible value.
The second function \code{\link{weight.standard}} uses the standard deviation
to assign weights. The more the values differ from one another within an attribute, the better the
assigned weight will be.

The \code{gamma} parameter measures the importance you want to give to the first
function. It acts also as a weight since the final weight vector is given by
\code{result = gamma * weight.highestValue + (1-gamma) * weight.standard}
}
\examples{
#Not runnable yet
weight.highAndStandard(myData, userid=10)
weight.highAndStandard(someData, 11, rounds="all")
weight.highAndStandard(laptop_data, 15, cost_ids=4, gamma = 0.3)
weight.highAndStandard(myData, 15, attr=1:4, "all", cost_ids=4, gamma = 0.75)

}

