% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeightFunctions.R
\name{weight.highestValue}
\alias{weight.highestValue}
\title{Calculates weights relative to the highest sum relative to other attributes}
\usage{
weight.highestValue(dataset, userid = NULL, attr = NULL, rounds = "all",
  cost_ids = NULL)
}
\arguments{
\item{dataset}{data.frame with the user generated data from a product
configurator. See \code{decisionMatrix} for specifications of the dataset.}

\item{userid}{a vector of integers that gives the information of which users
the matrix should be calculated. Vectorised.}

\item{attr}{attributes IDs, vector of integer numbers corresponding to the
  attributes you desire to use; attr are assumed to be 1-indexed. This
  function will calculate with all attributes and do the subsetting a
  posteriori.

  If you want to get the weights for only two attributes you will have to
  first use \code{\link{decisionMatrix}} and then pass it on to
  \code{\link{normalize.sum}} and \code{\link{entropy}}.}

\item{rounds}{integer vector or text option. Which steps of the configuration
process should be taken into account? Defaults are "all" in order to have
more data to calculate with. If \code{"first"} or \code{"last"} are entered
there will be only one rounds to gather data from, consequently all
attribtues will have the same weight.}

\item{cost_ids}{argument used to convert selected cost attributes into
benefit attributes. Integer vector. This functions uses an alternative
normalizing method \code{\link{normalize.highestValue}} that does not produce
negative values.}
}
\value{
a list of weight vector(s)
}
\description{
Awards higher weights to those attirbutes which consistently showed higher
values, measured against the highest posssible. Secondary function used for
main function \code{\link{weight.highAndStandard}}.
}

