% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetFunctions.R
\name{benefitToCostAttr}
\alias{benefitToCostAttr}
\title{HEader}
\usage{
benefitToCostAttr(dataset, aList, cost_ids = NULL)
}
\arguments{
\item{at}{}
}
\value{
Uni
}
\description{
HEader
}
\examples{
tbd

}
\seealso{
Other get.functions: \code{\link{getAllUserIds}},
  \code{\link{getAttrValues}},
  \code{\link{getDefaultRefps}},
  \code{\link{getRoundsById}}, \code{\link{getTableById}},
  \code{\link{get_attrs_ID}}, \code{\link{normalize}}
}

