% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetFunctions.R
\name{benefitToCostAttr}
\alias{benefitToCostAttr}
\title{HEader}
\usage{
benefitToCostAttr(dataset, aList, cost_ids = NULL)
}
\arguments{
\item{at}{}
}
\value{
Uni
}
\description{
HEader
}
\examples{
tbd

}
\seealso{
Other get functions: \code{\link{get_all_default_rps}},
  \code{\link{get_all_userids}},
  \code{\link{get_attr_values}},
  \code{\link{get_attrs_ID}},
  \code{\link{get_normalized_vec}},
  \code{\link{get_rounds_by_ID}},
  \code{\link{get_table_by_ID}}
}

