% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DecisionMatrix.R
\name{decisionMatrix}
\alias{decisionMatrix}
\title{DOCU New if only inputtin one vector in rounds, it will be converted into a list, if you need
different rounds for different userids, you need to provide them as a list of vectors.}
\usage{
decisionMatrix(dataset, userid = NULL, attr = NULL, rounds = NULL,
  cost_ids = NULL)
}
\description{
DOCU New if only inputtin one vector in rounds, it will be converted into a list, if you need
different rounds for different userids, you need to provide them as a list of vectors.
}

