% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetFunctions.R
\name{get_all_userids}
\alias{get_all_userids}
\title{Get all user IDs}
\usage{
get_all_userids(dataset)
}
\arguments{
\item{dataset}{the data.frame you want to input.}
}
\value{
a vector of integers with all user IDs. IDs are returned in the order
  they appear on the table.
}
\description{
Allows you to get all the user IDs in your data. It assumes you have a column
\code{$usid} that identifies each user's information of each round with its
corresponding user ID. Returns IDS unordered, i.e. in the order they appear
on the table.
}
\examples{
get_all_userids(example_data) #Example return: [1] 1 2 3 4 5  # ordered userids.
get_all_userids(data) #Example return: [1] 6 9 10 8 6 2 4  # returns IDs just as they are on the given table.

}
\references{
ProductConfig Github page:
  https://github.com/avilesd/productConfig
}
\seealso{
Other get functions: \code{\link{get_all_default_rps}},
  \code{\link{get_attr_values}},
  \code{\link{get_attrs_ID}},
  \code{\link{get_normalized_vec}},
  \code{\link{get_rounds_by_ID}},
  \code{\link{get_table_by_ID}}
}

