% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeightFunctions.R, R/dev_WeightFunctions.R
\name{get_attr_weight}
\alias{get_attr_weight}
\title{Weights calculating function}
\usage{
get_attr_weight(dataset = NULL, userid = NULL, weight = NULL,
  attr = NULL, rounds = NULL, cost_ids = NULL)

get_attr_weight(dataset = NULL, userid = NULL, weight = NULL,
  attr = NULL, rounds = NULL, cost_ids = NULL)
}
\arguments{
\item{dataset}{data.frame with the user generated data from a product
configurator. See Details for the specifications of the data.frame.}

\item{userid}{an integer that gives the information of which user you want
the data from: User ID.}

\item{weight}{integer vector. Represents the importance or relevance that an attribute has and the weight it should have in the
calculation of the prospect value.}

\item{attr}{attributes IDs, vector of integer numbers corresponding to the
attributes you desire to use; attr are assumed to be 1-indexed.}

\item{rounds}{integer vector. Which steps of the configuration process should be taken into account?}

\item{cost_ids}{argument used to convert selected cost attributes into
benefit attributes. Integer vector.}
}
\value{
Calculated weights according to the chosen weight_function
}
\description{
This function gives you an interface to all weight calculating functions in this package.

ndec11 <- norm.gainLoss(myData, 11, rounds="all")
}
\details{
Default value of FUN uses \code{\link{weight_higher_sum_value}} to get the \code{weight} vector. Below you can
find alternative functions implemented until now to calculate weights.

\code{attr} This function does handles different amount of attributes.

\code{cost_ids} Are passed along to the function you chosee (See Also), not handled here.

If you want to know more about the other parameters, look at \code{powerful_function}.
}
\examples{
get_attr_weight(pc_data, 11)
get_attr_weight(my_data, userid = 1000, cost_ids = 4)
get_attr_weight(monitor_data, 50)

}
\seealso{
Other weight functions: \code{\link{weight_higher_sum_value}}
}

