% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get_attrs_ID}
\alias{get_attrs_ID}
\title{Get the amount of attributes and their IDs}
\usage{
get_attrs_ID(x)
}
\arguments{
\item{x}{the data.frame you want to input.}
}
\value{
a vector of integers representing all existing attribute IDs.
}
\description{
Allows the user to know how many different attributes there are in the data.frame or table. The number of attributes comes directly
from the number of categories that a user can manipulate in a given product configurator. The function assumes that all attribute IDs are in a single column
with the name \code{$atid}. It also assumes, that all users interacted with the same amount of attributes.
}
\examples{
get_attrs_ID(my_data) #Example return: [1] 1 2 3 4 , meaning there are 4 attributes on my_data with ascending indexes.
get_attrs_ID(bad_data) #Warning: Attributes IDs should only be integers.
}
\references{
ProductConfig Github page: https://github.com/avilesd/productConfig
}
\seealso{
Other get functions: \code{\link{get_all_default_rps}};
  \code{\link{get_all_userids}};
  \code{\link{get_attr_values}};
  \code{\link{get_normalized_vec}};
  \code{\link{get_rounds_by_ID}};
  \code{\link{get_table_by_ID}}
}

