% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetFunctions.R
\name{get_normalized_vec}
\alias{get_normalized_vec}
\title{Normalize a vector}
\usage{
get_normalized_vec(num_vector)
}
\arguments{
\item{num_vector}{numeric vector you want to normalize.}
}
\value{
A normalized vector, i.e. anothernumeric vector with the sum of its
  absolute values equaling 1.
}
\description{
Divide all given vectors with the absolute sum of all the elements in the
vector, such that the sum of the result equals 1, that is if you take the
absolute value of the results. In other words, you can give negative values.
}
\examples{
get_normalized_vec(c(1,2,1,2)) #Returns: [1] 0.1666667 0.3333333 0.1666667 0.3333333
get_normalized_vec(c(1,2,1,-2)) #Returns: [1] 0.1666667 0.3333333 0.1666667 -0.3333333

}
\references{
ProductConfig Github page:
  https://github.com/avilesd/productConfig
}
\seealso{
Other get functions: \code{\link{get_all_default_rps}},
  \code{\link{get_all_userids}},
  \code{\link{get_attr_values}},
  \code{\link{get_attrs_ID}},
  \code{\link{get_rounds_by_ID}},
  \code{\link{get_table_by_ID}}
}

