% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetFunctions.R
\name{getDefaultRefps}
\alias{getDefaultRefps}
\title{Get Default Reference Points for selected user}
\usage{
getDefaultRefps(dataset, userid = NULL)
}
\arguments{
\item{dataset}{the data.frame you want to analyze. For the default values for
more or all users, see \code{\link{powerful_function}}}

\item{user_id}{an integer, that identifies which rows (from which user) of the
whole given dataset should be extracted.}
}
\value{
A numeric vector with column names representing the attribute ID each
 reference point belongs to.
}
\description{
With Default Referenece Points we mean the initial configuration of the
product configurator, that is the value for each attribute that was selected
as default before the user interacted with the configurator. It assumes there
is a column with the name 'round' from which to read the default values (in
round = 0) for each user. The  The default values of each 'category' can be
interpreted as possible Reference Points for the decision maker. See Source
for more about this assumption.
}
\examples{
get_all_default_rps(camera2_config_data, 100)
get_all_default_rps(as.data.frame(matrix_full), 55)
}
\seealso{
\code{\link{powerful_function}} allows you to run this and any other
 function in this package for more than one user.

Other get functions: \code{\link{benefitToCostAttr}},
  \code{\link{getAllUserIds}}, \code{\link{getAttrValues}},
  \code{\link{getRoundsById}}, \code{\link{getTableById}},
  \code{\link{get_attrs_ID}}, \code{\link{normalize}}
}

