% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetFunctions.R
\name{getTableById}
\alias{getTableById}
\title{Extract the part of the provided dataset that corresponds to a user id.}
\usage{
getTableById(dataset, userid = NULL, ...)
}
\arguments{
\item{x}{the data.frame (table) you want to separate in a smaller data.frame
for a specific user. For more users at once see
\code{\link{powerful_function}} or manually.}

\item{user_id}{an integer, that identifies which rows (from which user) of
the whole given dataset should be extracted.}
}
\value{
A data.frame with the same columns as the original, but only with
  rows about the provided \code{user_id}.
}
\description{
With user_id you select the rows from one user, all given columns of x remain
intact for that user. We assume the given dataset was generated by data from
a product configuration, and therefore has as one column the user id. Other
columns may include, but not limited to, attribute IDs, round, selected
attribute and selectable attribute.
}
\examples{
get_table_by_ID(camera2_config_data, 9)
get_table_by_ID(as.data.frame(matrix_full), 12)
}
\seealso{
\code{\link{powerful_function}} allows you to run this and any other
  function in this package for more than one user.

Other get functions: \code{\link{benefitToCostAttr}},
  \code{\link{getAllUserIds}}, \code{\link{getAttrValues}},
  \code{\link{getDefaultRefps}},
  \code{\link{getRoundsById}}, \code{\link{get_attrs_ID}},
  \code{\link{normalize}}
}

