% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PT-Functions.R
\name{lossFunction}
\alias{lossFunction}
\title{Calculates the loss for one attribute}
\usage{
lossFunction(v1, v2)
}
\arguments{
\item{s_ij}{value of attribute j in round i}

\item{e_j}{reference point off attribute j}
}
\value{
loss, numeric value.
}
\description{
A simple function that given the value of an attribute (s_ij) and the
reference point of the same attribute, calculates the loss and returns it. It
is not built as a stand alone function, rather as an object to be used by
other major functions, such as \code{gain_matrix loss_matrix}.
}
\details{
For understanding how this works, see the function itself or refer
to the paper. It handles only discrete numbers, so no interval numbers. Also
a point to improve further on.
}
\examples{
loss_fun_a(5, 1)  # returns: 0
loss_fun_a(2, 3)  # returns: -1

}

