% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PT-Functions.R
\name{lossMatrix}
\alias{lossMatrix}
\title{Loss matrix}
\usage{
lossMatrix(dataset, userid = NULL, attr = NULL, rounds = NULL,
  refps = NULL, cost_ids = NULL)
}
\arguments{
\item{refps}{a list of numeric vectors, one for each user. Reference Points:
each point corresponds to one attribute, therefore the amount of attributes
and of refps entered, should be equal. Default assumes the refps as the
default values of the initial product configuration for each user. You may
fully or partially enter your own reference points, check below for more
info.}
}
\value{
a list of loss matrices, one for each user.
}
\description{
Creates the Loss matrix parting from a decision matrix and a vector containing
the reference points (typically the status-quo). A loss represents a positive
difference between a given value in the decision matrix and its corresponding
reference point [1,2]. This functions is intended to use only for single
referene point theories, not for multiple reference point approaches. For the
latter, refer to \code{\link{overallDRP}}, and \code{\link{overallTRP}}.
}
\details{
The returned lossMatrix has: ncol = number of attributes you selected
 or all(default) and nrow= number of rounds you selected or the first and
 last(default) for a selected user. Results are unnamed.

 \code{dataset} We assume the input data.frame has following columns usid =
 User IDs, round = integers indicating which round the user is in (0-index
 works best for 'round'), atid = integer column for referring the attribute
 ID (1 indexed), selected = numeric value of the attribute for a specific,
 given round, selectable = amount of options the user can chose at a given
 round, with the current configuration. This is a necessary parameter.

 \code{userid} is a necessary parameter, without it you'll get a warning.
 Default is NULL.

 \code{attr} Default calculates with all attributes. Attributes are
 automatically read from provided table, it is important you always provide
 the complete dataset so that the package functions properly. Moreover the
 attributes will not be sorted. Output columns are returned in the ordered
 they were inputed.

 \code{rounds} If you need to compute different rounds for each user you
 enter, this argument accepts a list of integer vectors indicating which
 rounds should be used for each user. The function does not read names, it
 works in the order the list was given.

 \code{refps} If you only want to see the results for one attribute you may
 enter only a couple of reference points but you have to tell the function
 which attributes you want to use those referene points for. So the amount of
 attr and of refps should be the same. Moreover the functions always orders
 de attr, so be sure to input the reference point also in an ascending order
 corresponding to their attributes. (refps will not be ordered)

 \code{cost_ids} If \code{attr} and \code{cost_ids} differ, the functions
 will first compute the entire decision matrix using the \code{cost_ids} and
 only in the end will it 'subset' the result to the desired \code{attr}.
}
\examples{
#Not Runnable yet
lossMatrix(pc_config_data, 9:11)
lossMatrix(my_data, userid = 11, rounds="all")
lossMatrix(keyboard_data, 60, refps = c(1,3,4,0), cost_ids = 4)
lossMatrix(data1, 2, rounds = "last", attr = 1, cost_ids=1)
}
\references{
[1] Fan, Z. P., Zhang, X., Chen, F. D., & Liu, Y. (2013). Multiple
 attribute decision making considering aspiration-levels: A method based on
 prospect theory. Computers & Industrial Engineering, 65(2), 341-350.

 [2]Kahneman, D., & Tversky, A. (1979). Prospect theory: An analysis of
 decision under risk. Econometrica: Journal of the Econometric Society,
 263-291.
}

