% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProspectValueFunctions.R
\name{overallDRP}
\alias{overallDRP}
\title{Calcultes the overall prospect values using the DRP approach (dual reference
point)}
\usage{
overallDRP(dataset, userid = NULL, attr = NULL, rounds = NULL,
  cost_ids = NULL, weight = NULL, weightFUN = "differenceToIdeal",
  dual.refps = c(sq = NA, g = NA), lambda = 2.25, delta = 0.8,
  consumption_fun = NULL, gamma)
}
\arguments{
\item{dataset}{data.frame with the user generated data from a product
configurator. See \code{decisionMatrix} for specifications of the dataset.}

\item{userid}{a vector of integers that gives the information of which users
the matrix should be calculated. Vectorised.}

\item{attr}{attributes IDs, vector of integer numbers corresponding to the
attributes you desire to use; \code{attr} are assumed to be 1-indexed.}

\item{rounds}{integer vector or text option. Which steps of the configuration
process should be shown? Defaults are first and last step. Text options are
\code{all, first, last}.}

\item{cost_ids}{argument used to convert selected cost attributes into
benefit attributes. Integer vector.}

\item{weight}{numeric vector. Represents the importance or relevance that an
attribute has and the weight it should have in the calculation of the
prospect value.}

\item{weightFUN}{indicated which weight function should be used to calculate
the weight vector, the options are \code{"differenceToIdeal", "entropy",
"highAndStandard", "standard"}. Default ist \code{differencetoIdeal}.}

\item{dual.refps}{numeric - two numbers indicating the status-quo and the
aspiration level(goal) for the given attributes. Status-quo should always
be the first input.}

\item{lambda}{numeric - parameter of loss aversion for the value function as
given by reference[1]. Default value is 2.25 as in [2] and should be
\code{lambda > 1}.}

\item{delta}{numeric - expresses the relative importance of the aspiration
level to other factors. Default is 0.8 and it should satisfy \code{0 <
delta <1}.}

\item{consumption_fun}{non-working parameter for future developments. Leave
at NULL.}

\item{gamma}{numeric and between 0 and 1. It is a parameter used for the function
\code{\link{weight.highAndStandard}} and
\code{\link{highAndStandard}}, please the former to know its use.}
}
\value{
overall prospect values for each alternative/row
}
\description{
The overall prospect values for each alternative (round) are calculated using
two components in a method described in [3]: (1) the value matrix, which is
calculated using \code{\link{dual.valueMatrix}} as given by [1] and (2) the
decision weights, which can be calculated in this package using three
functions. See Details of weight attribute. Its equivalent function for
prospect theory and the tri-reference point theory are
\code{\link{overallPV}} and \code{\link{overallTRP}}.
}
\examples{
#Not runnable yet
overallDRP(pc_config_data, 9, dual.refps = c(1, 3.5))
overallDRP(pc_config_data, 9, dual.refps = c(1, 3.5), lambda=2, delta=0.5)
overallDRP(aDataFrame, userid = 100, rounds = "all", dual.refps = c(1, 2))
overallDRP(myData, userid = 11, attr =  1, cost_ids = 1, dual.refps = c(8, 2))
overallDRP(full_data, 15:16 ,attr = c(1,2,3), rounds="all", dual.refps=matrix(1:6, 3, 2, byrow=T))

}
\references{
[1] Golman, R., & Loewenstein, G. (2011). Explaining Nonconvex
  Preferences with Aspirational and Status Quo Reference Dependence. Mimeo,
  Carnegie Mellon University.

  [2] Tversky, A., & Kahneman, D. (1992). Advances in prospect theory:
  Cumulative representation of uncertainty. Journal of Risk and uncertainty,
  5(4), 297-323.

  [3] Fan, Z. P., Zhang, X., Chen, F. D., & Liu, Y. (2013). Multiple
  attribute decision making considering aspiration-levels: A method based on
  prospect theory. Computers & Industrial Engineering, 65(2), 341-350.
}
\seealso{
\code{\link{decisionMatrix, overallTRP, overallDRP,
  weight.differenceToIdeal, weight.entropy, weight.highAndStandard}}
}

