% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/powerful_functions.R
\name{powerful_function}
\alias{powerful_function}
\title{Calculate the Gain matrix}
\usage{
powerful_function(dataset, userid = NULL, FUN = decision_matrix,
  attr = NULL, rounds = NULL, refps = NULL, cost_ids = NULL,
  weight = NULL, alpha = 0.88, beta = 0.88, lambda = 2.25,
  gainm = TRUE)
}
\arguments{
\item{dataset}{}

\item{userid}{}

\item{attr}{}

\item{rounds}{}
}
\value{
(NOT CHANGED)A decision matrix with j amount of columns and i amount of rows. Colnames = attrIDs and rownames = chosen rounds.\code{user_id}.
}
\description{
Creates the Gains matrix parting from a decision matrix and a vector containing the reference points (aspiration-levels).
}
\details{
(NOT CHANGED)Creates a decision matrix, which is necessary to create the 'Gains' and 'Losses' matrices after. If no attribute vector is given,
(NOT CHANGED)containing the attributes IDs and if no alternatives are passed on as rounds, to be considered in the decision matrix, the function
(NOT CHANGED)will go with the defaults and extract them from the dataset. Will be used only for one user (one userid).
}
\examples{
decision_matrix(camera2_config_data, 11)
decision_matrix(my_data, userid = 11, attr = c(1,3,5))
decision_matrix(another_data, userid = 80, rounds = c(1,2,3,7,8,9))
decision_matrix(data2, 2, rounds = "all")
}

