% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PT-Functions.R
\name{prospect_value_matrix_extend}
\alias{prospect_value_matrix_extend}
\title{Calculate Value Matrix}
\usage{
prospect_value_matrix_extend(ngain = NULL, nloss = NULL, alpha = 0.88,
  beta = 0.88, lambda = 2.25)
}
\arguments{
\item{ngain}{normalized gain matrix}

\item{nloss}{normalized loss matrix}

\item{alpha}{numeric between [0, 1]. Determines the concativity of the value
function as given by the value function[1].}

\item{beta}{numeric between [0, 1]. Determines the convexity of the value
function as given by the value function[1]}

\item{lambda}{lambda > 1. Parameter of loss aversion for the value function
as given by the value function[1].}
}
\value{
the value matrix
}
\description{
Given both normalized gain and loss matrices, this function calculates the
value matrix with the value function from Prospect Theory (Reference[1]). It
differs from other functions in this package in that it does not take all
parameters into account. Yo need to pre-calculate the matrices. See Details.
}
\details{
You need to pre-calculate the normalized gain and loss matrices, for
  example with \code{\link{norm_g_l_matrices}} and give them as a parameter.
  This is one of the few functions of this package that do not allow you to
  give the raw data from your product Configurator, but rather calculate a
  previous result.
}
\examples{
prospect_value_matrix_extend(my_ngain, my_nloss)
prospect_value_matrix_extend(ngain = matrix1, nloss = matrix2, alpha = 0.95)

}
\references{
[1] Kahneman, D., & Tversky, A. (1979). Prospect theory: An
  analysis of decision under risk. Econometrica: Journal of the Econometric
  Society, 263-291.

  [2] Fan, Z. P., Zhang, X., Chen, F. D., & Liu, Y. (2013). Multiple
  attribute decision making considering aspiration-levels: A method based on
  prospect theory. Computers & Industrial Engineering, 65(2), 341-350.
}
\seealso{
\code{\link{pvMatrix}}
}

