% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRPFunctions.R
\name{trpValueMatrix}
\alias{trpValueMatrix}
\title{DOCU: you can enter cost_ids normally, program will recognize for which attr it should use the cost_ids}
\usage{
trpValueMatrix(dataset, userid = NULL, attr = NULL, rounds = NULL,
  cost_ids = NULL, mr = 0.5, sq = 1.5, g = 2.5, beta_f = 5,
  beta_l = 1, beta_g = 1, beta_s = 3)
}
\description{
DOCU: you can enter cost_ids normally, program will recognize for which attr it should use the cost_ids
}

